requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiohttp==3.7.4.post0
    # via discord-py
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
async-timeout==3.0.1
    # via aiohttp
asyncio==3.4.3
    # via -r requirements.in
attrs==21.4.0
    # via
    #   aiohttp
    #   outcome
    #   trio
beautifulsoup4==4.11.1
    # via -r requirements.in
certifi==2022.6.15
    # via
    #   requests
    #   urllib3
cffi==1.15.1
    # via
    #   cryptography
    #   trio
chardet==4.0.0
    # via aiohttp
charset-normalizer==2.1.0
    # via requests
cryptography==37.0.2
    # via
    #   pyopenssl
    #   urllib3
discord==1.7.3
    # via -r requirements.in
discord-py==1.7.3
    # via discord
h11==0.13.0
    # via wsproto
idna==3.3
    # via
    #   requests
    #   trio
    #   urllib3
    #   yarl
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
outcome==1.2.0
    # via trio
pathlib==1.0.1
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyopenssl==22.0.0
    # via urllib3
pysocks==1.7.1
    # via urllib3
python-dateutil==2.8.2
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
selenium==4.3.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.3.2.post1
    # via beautifulsoup4
trio==0.21.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
typing-extensions==4.3.0
    # via aiohttp
urllib3[secure,socks]==1.26.9
    # via
    #   requests
    #   selenium
wsproto==1.1.0
    # via trio-websocket
yarl==1.7.2
    # via aiohttp
